requirements.txt
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.19.0
Pillow>=8.0.0
efficientnet-pytorch>=0.7.1
einops>=0.3.0
python_dotenv==1.1.1
